+++ sdk/privatedns/Azure.ResourceManager.PrivateDns/src/Customization/Models/PrivateDnsCnameRecordData.Serialization.cs
@@ -3,6 +3,8 @@
 
 #nullable disable
 
+using System;
+using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Text.Json;
 using Azure;
@@ -12,21 +14,28 @@ using Azure.ResourceManager.PrivateDns.Models;
 
 namespace Azure.ResourceManager.PrivateDns
 {
-    public partial class PrivateDnsCnameRecordData : IUtf8JsonSerializable
+    public partial class PrivateDnsCnameRecordData : IUtf8JsonSerializable, IJsonModel<PrivateDnsCnameRecordData>
     {
-        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
+        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer) => ((IJsonModel<PrivateDnsCnameRecordData>)this).Write(writer, new ModelReaderWriterOptions("W"));
+
+        void IJsonModel<PrivateDnsCnameRecordData>.Write(Utf8JsonWriter writer, ModelReaderWriterOptions options)
         {
+            var format = options.Format == "W" ? ((IPersistableModel<PrivateDnsCnameRecordData>)this).GetFormatFromOptions(options) : options.Format;
+            if (format != "J")
+            {
+                throw new FormatException($"The model {nameof(PrivateDnsCnameRecordData)} does not support '{format}' format.");
+            }
             writer.WriteStartObject();
             if (Optional.IsDefined(ETag))
             {
-                writer.WritePropertyName("etag");
+                writer.WritePropertyName("etag"u8);
                 writer.WriteStringValue(ETag.Value.ToString());
             }
-            writer.WritePropertyName("properties");
+            writer.WritePropertyName("properties"u8);
             writer.WriteStartObject();
             if (Optional.IsCollectionDefined(Metadata))
             {
-                writer.WritePropertyName("metadata");
+                writer.WritePropertyName("metadata"u8);
                 writer.WriteStartObject();
                 foreach (var item in Metadata)
                 {
@@ -37,20 +46,53 @@ namespace Azure.ResourceManager.PrivateDns
             }
             if (Optional.IsDefined(TtlInSeconds))
             {
-                writer.WritePropertyName("ttl");
+                writer.WritePropertyName("ttl"u8);
                 writer.WriteNumberValue(TtlInSeconds.Value);
             }
             if (Optional.IsDefined(PrivateDnsCnameRecord))
             {
-                writer.WritePropertyName("cnameRecord");
+                writer.WritePropertyName("cnameRecord"u8);
                 writer.WriteObjectValue(PrivateDnsCnameRecord);
             }
             writer.WriteEndObject();
+            if (options.Format != "W" && _serializedAdditionalRawData != null)
+            {
+                foreach (var item in _serializedAdditionalRawData)
+                {
+                    writer.WritePropertyName(item.Key);
+#if NET6_0_OR_GREATER
+				writer.WriteRawValue(item.Value);
+#else
+                    using (JsonDocument document = JsonDocument.Parse(item.Value))
+                    {
+                        JsonSerializer.Serialize(writer, document.RootElement);
+                    }
+#endif
+                }
+            }
             writer.WriteEndObject();
         }
 
-        internal static PrivateDnsCnameRecordData DeserializePrivateDnsCnameRecordData(JsonElement element)
+        PrivateDnsCnameRecordData IJsonModel<PrivateDnsCnameRecordData>.Create(ref Utf8JsonReader reader, ModelReaderWriterOptions options)
         {
+            var format = options.Format == "W" ? ((IPersistableModel<PrivateDnsCnameRecordData>)this).GetFormatFromOptions(options) : options.Format;
+            if (format != "J")
+            {
+                throw new FormatException($"The model {nameof(PrivateDnsCnameRecordData)} does not support '{format}' format.");
+            }
+
+            using JsonDocument document = JsonDocument.ParseValue(ref reader);
+            return DeserializePrivateDnsCnameRecordData(document.RootElement, options);
+        }
+
+        internal static PrivateDnsCnameRecordData DeserializePrivateDnsCnameRecordData(JsonElement element, ModelReaderWriterOptions options = null)
+        {
+            options ??= new ModelReaderWriterOptions("W");
+
+            if (element.ValueKind == JsonValueKind.Null)
+            {
+                return null;
+            }
             Optional<ETag> etag = default;
             ResourceIdentifier id = default;
             string name = default;
@@ -61,9 +103,11 @@ namespace Azure.ResourceManager.PrivateDns
             Optional<string> fqdn = default;
             Optional<bool> isAutoRegistered = default;
             Optional<PrivateDnsCnameRecordInfo> cnameRecordInfo = default;
+            IDictionary<string, BinaryData> serializedAdditionalRawData = default;
+            Dictionary<string, BinaryData> additionalPropertiesDictionary = new Dictionary<string, BinaryData>();
             foreach (var property in element.EnumerateObject())
             {
-                if (property.NameEquals("etag"))
+                if (property.NameEquals("etag"u8))
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
@@ -73,22 +117,22 @@ namespace Azure.ResourceManager.PrivateDns
                     etag = new ETag(property.Value.GetString());
                     continue;
                 }
-                if (property.NameEquals("id"))
+                if (property.NameEquals("id"u8))
                 {
                     id = new ResourceIdentifier(property.Value.GetString());
                     continue;
                 }
-                if (property.NameEquals("name"))
+                if (property.NameEquals("name"u8))
                 {
                     name = property.Value.GetString();
                     continue;
                 }
-                if (property.NameEquals("type"))
+                if (property.NameEquals("type"u8))
                 {
                     type = new ResourceType(property.Value.GetString());
                     continue;
                 }
-                if (property.NameEquals("systemData"))
+                if (property.NameEquals("systemData"u8))
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
@@ -98,7 +142,7 @@ namespace Azure.ResourceManager.PrivateDns
                     systemData = JsonSerializer.Deserialize<SystemData>(property.Value.ToString());
                     continue;
                 }
-                if (property.NameEquals("properties"))
+                if (property.NameEquals("properties"u8))
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
@@ -107,7 +151,7 @@ namespace Azure.ResourceManager.PrivateDns
                     }
                     foreach (var property0 in property.Value.EnumerateObject())
                     {
-                        if (property0.NameEquals("metadata"))
+                        if (property0.NameEquals("metadata"u8))
                         {
                             if (property0.Value.ValueKind == JsonValueKind.Null)
                             {
@@ -122,7 +166,7 @@ namespace Azure.ResourceManager.PrivateDns
                             metadata = dictionary;
                             continue;
                         }
-                        if (property0.NameEquals("ttl"))
+                        if (property0.NameEquals("ttl"u8))
                         {
                             if (property0.Value.ValueKind == JsonValueKind.Null)
                             {
@@ -132,12 +176,12 @@ namespace Azure.ResourceManager.PrivateDns
                             ttl = property0.Value.GetInt64();
                             continue;
                         }
-                        if (property0.NameEquals("fqdn"))
+                        if (property0.NameEquals("fqdn"u8))
                         {
                             fqdn = property0.Value.GetString();
                             continue;
                         }
-                        if (property0.NameEquals("isAutoRegistered"))
+                        if (property0.NameEquals("isAutoRegistered"u8))
                         {
                             if (property0.Value.ValueKind == JsonValueKind.Null)
                             {
@@ -147,7 +191,7 @@ namespace Azure.ResourceManager.PrivateDns
                             isAutoRegistered = property0.Value.GetBoolean();
                             continue;
                         }
-                        if (property0.NameEquals("cnameRecord"))
+                        if (property0.NameEquals("cnameRecord"u8))
                         {
                             if (property0.Value.ValueKind == JsonValueKind.Null)
                             {
@@ -160,8 +204,44 @@ namespace Azure.ResourceManager.PrivateDns
                     }
                     continue;
                 }
+                if (options.Format != "W")
+                {
+                    additionalPropertiesDictionary.Add(property.Name, BinaryData.FromString(property.Value.GetRawText()));
+                }
             }
-            return new PrivateDnsCnameRecordData(id, name, type, systemData.Value, Optional.ToNullable(etag), Optional.ToDictionary(metadata), Optional.ToNullable(ttl), fqdn.Value, Optional.ToNullable(isAutoRegistered), cnameRecordInfo.Value);
+            serializedAdditionalRawData = additionalPropertiesDictionary;
+            return new PrivateDnsCnameRecordData(id, name, type, systemData.Value, Optional.ToNullable(etag), Optional.ToDictionary(metadata), Optional.ToNullable(ttl), fqdn.Value, Optional.ToNullable(isAutoRegistered), cnameRecordInfo.Value, serializedAdditionalRawData);
         }
+
+        BinaryData IPersistableModel<PrivateDnsCnameRecordData>.Write(ModelReaderWriterOptions options)
+        {
+            var format = options.Format == "W" ? ((IPersistableModel<PrivateDnsCnameRecordData>)this).GetFormatFromOptions(options) : options.Format;
+
+            switch (format)
+            {
+                case "J":
+                    return ModelReaderWriter.Write(this, options);
+                default:
+                    throw new FormatException($"The model {nameof(PrivateDnsCnameRecordData)} does not support '{options.Format}' format.");
+            }
+        }
+
+        PrivateDnsCnameRecordData IPersistableModel<PrivateDnsCnameRecordData>.Create(BinaryData data, ModelReaderWriterOptions options)
+        {
+            var format = options.Format == "W" ? ((IPersistableModel<PrivateDnsCnameRecordData>)this).GetFormatFromOptions(options) : options.Format;
+
+            switch (format)
+            {
+                case "J":
+                    {
+                        using JsonDocument document = JsonDocument.Parse(data);
+                        return DeserializePrivateDnsCnameRecordData(document.RootElement, options);
+                    }
+                default:
+                    throw new FormatException($"The model {nameof(PrivateDnsCnameRecordData)} does not support '{options.Format}' format.");
+            }
+        }
+
+        string IPersistableModel<PrivateDnsCnameRecordData>.GetFormatFromOptions(ModelReaderWriterOptions options) => "J";
     }
 }