+++ sdk/hybridaks/Azure.ResourceManager.HybridContainerService/CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Features Added
 
 - Upgraded api-version tag from 'package-preview-2023-11' to 'package-2024-01'. Tag detail available at https://github.com/Azure/azure-rest-api-specs/blob/8e674dd2a88ae73868c6fa7593a0ba4371e45991/specification/hybridaks/resource-manager/readme.md.
+- Enable the new model serialization by using the System.ClientModel, refer this [document](https://aka.ms/azsdk/net/mrw) for more details.
 
 ## 1.0.0-beta.4 (2024-01-03)