+++ sdk/hybridaks/Azure.ResourceManager.HybridContainerService/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 This release is the first stable release of the Hybrid Container Service Management library.
 
+### Other Changes
+
+- Upgraded dependent `Azure.ResourceManager` to 1.10.1.
+
 ## 1.0.0-beta.5 (2024-01-15)
 
 ### Features Added