+++ sdk/hybridaks/Azure.ResourceManager.HybridContainerService/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.6 (Unreleased)
+## 1.1.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,14 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0 (2024-01-26)
+
+This release is the first stable release of the Hybrid Container Service Management library.
+
+### Other Changes
+
+- Upgraded dependent `Azure.ResourceManager` to 1.10.1.
 
 ## 1.0.0-beta.5 (2024-01-15)